requirements.txt
# Core ML/AI libraries
torch>=1.11.0
transformers>=4.20.0
datasets>=2.0.0
tokenizers>=0.12.0

# Data processing
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0

# Translation-specific
sacremoses>=0.0.53
sentencepiece>=0.1.96

# API and web framework
fastapi>=0.78.0
uvicorn>=0.17.0
streamlit>=1.10.0

# Utilities
tqdm>=4.64.0
matplotlib>=3.5.0
seaborn>=0.11.0
requests>=2.28.0

# Development and testing
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0

# Optional: For advanced preprocessing
nltk>=3.7
spacy>=3.4.0

# Configuration management
python-dotenv>=0.20.0
pyyaml>=6.0